+++ src/core/cpu/src/cpu_restrdis_force.cpp
@@ -23,13 +23,13 @@ void calc_restrdis_forces() {
         i = restrdists[ir].ai - 1;
         j = restrdists[ir].aj - 1;
 
-        const double dx = static_cast<double>(coords[j].x) - static_cast<double>(coords[i].x);
-        const double dy = static_cast<double>(coords[j].y) - static_cast<double>(coords[i].y);
-        const double dz = static_cast<double>(coords[j].z) - static_cast<double>(coords[i].z);
+        const double dx = coords[j].x - coords[i].x;
+        const double dy = coords[j].y - coords[i].y;
+        const double dz = coords[j].z - coords[i].z;
 
         double lambda;
         if (restrdists[ir].ipsi != 0) {
-            lambda = static_cast<double>(lambdas[state]);
+            lambda = lambdas[state];
         } else {
             lambda = 1.0;
         }
